Source code adapted for GATE 8.6
