+++ creole.xml
@@ -1,7 +1,3 @@
-<!-- resource.xml LevenshteinGazeteer -->
-<!--  Ceilican, 4/7/2006 -->
-<!-- $Id: resource.xml,v 1.1 2001/05/31 16:06:02 hamish Exp $ -->
-
 <CREOLE-DIRECTORY>
 
 <CREOLE>
