--- lists.def
@@ -1,1 +1,0 @@
-city.lst:location:city
