--- nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=45349a66
-build.xml.script.CRC32=8a8b905b
-build.xml.stylesheet.CRC32=e2435cb5
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=45349a66
-nbproject/build-impl.xml.script.CRC32=662118d4
-nbproject/build-impl.xml.stylesheet.CRC32=3b89b5a6
