--- nbproject/private/private.properties
@@ -1,5 +1,0 @@
-application.args=
-javac.debug=true
-javadoc.preview=true
-jaxws.endorsed.dir=C:\\Program Files\\NetBeans\\ide7\\modules\\ext\\jaxws21\\api
-user.properties.file=C:\\Users\\Bruno\\.netbeans\\5.5.1\\build.properties
