+++ pom.xml
@@ -1,0 +1,63 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <groupId>bwp.gate.bwpgazetter</groupId>
+    <artifactId>bwpgazetteer</artifactId>
+    <version>1.0</version>
+
+    <packaging>jar</packaging>
+    <name>bwp-gazetteer</name>
+
+    <description>BWP Gazetteer</description>
+
+    <properties>
+
+        <skipTests>true</skipTests>
+        <json.version>1.1.1</json.version>
+        <java.version>1.8</java.version>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+
+    </properties>
+
+    <dependencies>
+
+        <!-- https://mvnrepository.com/artifact/uk.ac.gate/gate-core -->
+        <dependency>
+            <groupId>uk.ac.gate</groupId>
+            <artifactId>gate-core</artifactId>
+            <version>8.6</version>
+        </dependency>
+
+        <!-- https://mvnrepository.com/artifact/uk.ac.gate.plugins/annie -->
+        <dependency>
+            <groupId>uk.ac.gate.plugins</groupId>
+            <artifactId>annie</artifactId>
+            <version>8.6</version>
+<!--            <scope>provided</scope>-->
+        </dependency>
+        <!-- https://mvnrepository.com/artifact/uk.ac.gate.plugins/tools -->
+        <dependency>
+            <groupId>uk.ac.gate.plugins</groupId>
+            <artifactId>tools</artifactId>
+            <version>8.6</version>
+        </dependency>
+
+    </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <configuration>
+                    <source>8</source>
+                    <target>8</target>
+                </configuration>
+            </plugin>
+
+        </plugins>
+    </build>
+</project>
